+++ src/skeletor/Food/Pancake.java
@@ -3,6 +3,8 @@ package skeletor.Food;
 import skeletor.Enums.E_Dodatki;
 import skeletor.Enums.E_KategoriaPosiłku;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Jarek on 2016-12-02.
  */
@@ -20,7 +22,7 @@ public class Pancake extends Dessert {
      * @param additives - dodatki do naleśników
      */
     public Pancake(String name, String[] components,
-                   double price, float weight,
+                   BigDecimal price, float weight,
                    E_KategoriaPosiłku category,
                    long preparation_time, E_Dodatki additives) {
         super(name, components, price, weight,
