+++ src/skeletor/Food/Salami.java
@@ -2,6 +2,8 @@ package skeletor.Food;
 
 import skeletor.Enums.E_KategoriaPosiłku;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Jarek on 2016-12-02.
  */
@@ -18,7 +20,7 @@ public class Salami extends Pizza {
      * @param size_big
      */
     public Salami(String name, String[] components,
-                  double price, float weight,
+                  BigDecimal price, float weight,
                   E_KategoriaPosiłku category,
                   long preparation_time, boolean size_big) {
         super(name, components, price, weight,
